% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_db.R
\name{download_db}
\alias{download_db}
\title{Download background data}
\usage{
download_db(pset, tset, nset, saveDir = tempdir(), verbose = FALSE)
}
\arguments{
\item{pset}{The Pharmacogenomic dataset name. View(pharmaco_sets) for currently available datasets.}

\item{tset}{The TCGA cancer type name. View(TCGA_sets) for currently available cancer types.}

\item{nset}{The normal tissue dataset name. Currently available datasets: GTEX.}

\item{saveDir}{A character(1) string with the folder path where the background data should be saved. Defaults to tempdir(). Will create directory if it does not exist.}

\item{verbose}{bool Should status messages be printed during download. Defaults to TRUE.}
}
\value{

}
\description{
Download background data
}
