% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_cor_map.R
\name{draw_cor_map}
\alias{draw_cor_map}
\title{Visualize drug effectiveness test result}
\usage{
draw_cor_map(mysRGES, topline, cell_info)
}
\arguments{
\item{mysRGES}{sRGES data.frame produced by get_reverse_score.}

\item{topline}{inidividual-related cell lines.}

\item{cell_info}{cell line information list produced by get_cell.}
}
\description{
Visualize drug effectiveness test result
}
