% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugcorTest.R
\name{drugcorTest}
\alias{drugcorTest}
\title{Evaluate drug effectiveness on relevant cell line}
\usage{
drugcorTest(mysRGES, topline, cell_info)
}
\arguments{
\item{mysRGES}{sRGES data.frame produced by get_reverse_score.}

\item{topline}{inidividual-related cell lines.}

\item{cell_info}{cell line information list produced by get_cell.}
}
\value{
A correlation list
}
\description{
Evaluate drug effectiveness on relevant cell line
}
