% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell.R
\name{get_cell}
\alias{get_cell}
\title{Compute individual-related cell line}
\usage{
get_cell(cmat, db.path = tempdir(), removeBatchEffect = TRUE, orgDB)
}
\arguments{
\item{cmat}{The RNA-seq count expression profiles of clinical query samples.}

\item{db.path}{A character(1) string with the folder path where the background data was saved.}

\item{removeBatchEffect}{Remove batch effects from expression data using limma (TRUE by default).}

\item{orgDB}{The human annotation db for ID convert.}
}
\value{
A cell line information list.
}
\description{
Compute individual-related cell line
}
