% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diff.R
\name{get_diff}
\alias{get_diff}
\title{Differential expression analysis}
\usage{
get_diff(
  data = NULL,
  DE_method = "edgeR",
  normalize_samples = TRUE,
  threshold_log2foldchange = 2.5,
  threshold_pval = 0.05,
  threshold_adjpval = 0.001,
  k = 1,
  n_topGenes = 500,
  parallel_cores = 2
)
}
\arguments{
\item{data}{The purified profiles of subgoups and control groups for each clinical query sample}

\item{DE_method}{edgeR, DESeq2 or limma DE analysis.}

\item{normalize_samples}{if TRUE, RUVSeq normalization is applied to either EdgeR or DESeq. No normalization needed for limma+voom.}

\item{threshold_log2foldchange}{threshold for log2foldchange. 2.5 by default.}

\item{threshold_pval}{threshold for pvalue. 0.05 by default.}

\item{threshold_adjpval}{threshold for adjusted pvalue. 0.001 by default.}

\item{k}{either k=1 (by default), k=2 or k=3, number of factors used in model matrix construction in RUVSeq normalization if normalize_samples=TRUE.}

\item{n_topGenes}{number of empiricaly differentially expressed genes estimated for RUVSeq normalization. Default is 5000.}

\item{parallel_cores}{number of cores to be used for parallel computing in DESeq2.}
}
\value{
A list with list of differentially expressed genes.
}
\description{
Differential expression analysis
}
