% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimate_score.R
\name{get_estimateScore}
\alias{get_estimateScore}
\title{Evaluate non-cancer cell infiltration in each subtype}
\usage{
get_estimateScore(mat, subtype, gs = NULL, doplot = TRUE)
}
\arguments{
\item{mat}{Profiles of selected TCGA data set.}

\item{subtype}{The NMF subtyping result.}

\item{gs}{A data.frame of 3 columns with Metagene, Cell_type and Annot.}

\item{doplot}{A logical value. If TRUE, draw the heatmap of non-cancer cell infiltration.}
}
\value{
A data.frame of GSVA enrichment scores.
}
\description{
Evaluate non-cancer cell infiltration in each subtype
}
