% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NMF.R
\name{get_NMF}
\alias{get_NMF}
\title{Implement transcriptome subtyping by NMF method}
\usage{
get_NMF(
  mat,
  method = "MAD",
  value = 1500,
  clusterNum = NULL,
  rank = 2:6,
  nrun = 30,
  seed = NULL,
  doPlot = TRUE
)
}
\arguments{
\item{mat}{Profiles of selected TCGA data set.}

\item{method}{Method for feature selection, optional 'MAD', 'VAR' and 'PCA' ('MAD' by default).}

\item{value}{A numeric value for biological feature selection.
If method='MAD' or 'VAR', the top number of value features are selected.
If method='PCA', the value of principal component is selected.}

\item{clusterNum}{Number of subtypes for the data set.}

\item{rank}{A numeric vector for the estimation of clusterNum. It will be overrided if clusterNum is not NULL.}

\item{nrun}{Number of runs to perform NMF. A default of 30 runs are performed, allowing the computation of a consensus matrix that is used in selecting the best result for cancer subtypes identification as Consensus Clustering method.}

\item{seed}{A numeric value is used to seed the random number generator before generating a random starting point.}

\item{doPlot}{A logical value. If true, draw the heatmap for the distance matrix with samples ordered to form clusters.}
}
\value{
The NMF subtyping result.
}
\description{
Implement transcriptome subtyping by NMF method
}
