% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reverse_score.R
\name{get_reverse_score}
\alias{get_reverse_score}
\title{Compute sRGES}
\usage{
get_reverse_score(
  dz_signature = NULL,
  max_gene_size = 500,
  LINCS_data = NULL,
  permutations = 10000
)
}
\arguments{
\item{dz_signature}{disease signature. Make sure input data frame has a gene Symbol column, otherwise an error is produced. It must be an UPPERCASE gene symbol.}

\item{max_gene_size}{maximum number of disease genes used for drug prediction. By default 50 for each side (up/down).}

\item{LINCS_data}{LINCS data. By default set to NULL, which means use LINCS_978.}

\item{permutations}{number of permutations, by default 10000.}
}
\value{
A data.frame containing scores and p.values for every instance. data.frame contains drug id in pert_iname collumn, n contains the number of instances for this drug, mean, median and sd of sRGES RGES sores.
}
\description{
Compute sRGES
}
