% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subgroup.R
\name{get_subgroup}
\alias{get_subgroup}
\title{Identify subgroup and control group for each clinical query sample}
\usage{
get_subgroup(
  cmat,
  mat,
  subtype,
  k = 10,
  biopsy = NULL,
  adjacent = FALSE,
  doplot = TRUE,
  db.path = tempdir()
)
}
\arguments{
\item{cmat}{The RNA-seq count expression profiles of clinical query samples.}

\item{mat}{Profiles of selected TCGA data set.}

\item{subtype}{The NMF subtyping result.}

\item{k}{The subgroup size (10 by default. Optional 3, 5, 10.)}

\item{biopsy}{Tissue sources of the control group (Use View(GTEX_set) for optional 51 biopsy.sites. By default set to NULL, which will select control group based on distance between TCGA and GTEX samples.)}

\item{adjacent}{A logical value. By default set to FALSE. If TRUE, only tissue with sample.type 'adjacent' from GTEX_set would be used instead of 'normal'.}

\item{doplot}{A logical value. If TRUE, draw the heatmap of Spearman correlation coefficients between query samples and subtypes.}

\item{db.path}{A character(1) string with the folder path where the background data was saved.}
}
\value{
Profiles of subgoups and control groups for each clinical query sample.
}
\description{
Identify subgroup and control group for each clinical query sample
}
