% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_db.R
\name{select_db}
\alias{select_db}
\title{Preprocess transcriptome profiling and filter TCGA samples}
\usage{
select_db(
  Assay,
  cmat,
  MSI_status = NULL,
  driver_gene = NULL,
  MUT_status = NULL,
  CNA_status = NULL,
  geneList = NULL,
  minSampleSize = 10,
  removeBatchEffect = TRUE,
  OrgDb
)
}
\arguments{
\item{Assay}{The TCGA data set of selected cancer type.}

\item{cmat}{The RNA-seq count expression profiles of clinical query samples.}

\item{MSI_status}{Filter TCGA samples with the MSI status (MSI or MSS).}

\item{driver_gene}{A character with the driver gene symbol.}

\item{MUT_status}{Filter TCGA samples with the MUT status of driver gene (MUT or NULL).}

\item{CNA_status}{Filter TCGA samples with the CNA status of driver gene (AMP, GAIN, HETLOSS, HOMDEL).}

\item{geneList}{A data.frame of 2 column with gene and status. Only used in multiple driver genes.}

\item{minSampleSize}{Minimal size of the outpputted samples (10 by default).}

\item{removeBatchEffect}{Remove batch effects from expression data using limma (TRUE by default).}

\item{OrgDb}{The human annotation db for ID convert.}
}
\value{
A large list with processed profiles of clinical query samples (cmat), and processed and filtered profiles of selected TCGA data set (mat).
}
\description{
Preprocess transcriptome profiling and filter TCGA samples
}
